[REFACTOR/#133] 요청, 응답 데이터 model 패키지로 변경
