+++ app/src/main/java/com/example/teumteum/data/remote/onboarding/model/NicknameJobResponse.kt
@@ -1,9 +1,9 @@
-package com.example.teumteum.data.remote.onboarding.dto
+package com.example.teumteum.data.remote.onboarding.model
 
 import com.google.gson.annotations.SerializedName
 
-data class ProfileImageResponse(
-    @SerializedName("isSuccess") val isSuccess: Boolean,
+data class NicknameJobResponse(
+    @SerializedName("isSuccess") val isSuccess: String,
     @SerializedName("code") val code: String,
     @SerializedName("message") val message: String,
 )
