+++ app/src/main/java/com/example/teumteum/data/remote/onboarding/model/PresignedResponse.kt
@@ -1,0 +1,10 @@
+package com.example.teumteum.data.remote.onboarding.model
+
+import com.google.gson.annotations.SerializedName
+
+data class PresignedResponse(
+    @SerializedName("presignedUrl") val presignedUrl: String,
+    @SerializedName("fileName") val fileName: String
+)
+
+
