+++ app/src/main/java/com/example/teumteum/data/remote/onboarding/model/ProfileImageResponse.kt
@@ -1,9 +1,9 @@
-package com.example.teumteum.data.remote.onboarding.dto
+package com.example.teumteum.data.remote.onboarding.model
 
 import com.google.gson.annotations.SerializedName
 
-data class NicknameJobResponse(
-    @SerializedName("isSuccess") val isSuccess: String,
+data class ProfileImageResponse(
+    @SerializedName("isSuccess") val isSuccess: Boolean,
     @SerializedName("code") val code: String,
     @SerializedName("message") val message: String,
 )
