+++ app/src/main/java/com/example/teumteum/data/remote/onboarding/model/ScheduleRequest.kt
@@ -1,7 +1,6 @@
-package com.example.teumteum.data.remote.onboarding.dto
+package com.example.teumteum.data.remote.onboarding.model
 
 import com.google.gson.annotations.SerializedName
-import java.time.LocalTime
 
 data class ScheduleRequest(
     @SerializedName("routine") val routine: List<Schedule>
