+++ app/src/main/java/com/example/teumteum/data/remote/onboarding/model/SleepPatternResponse.kt
@@ -1,4 +1,4 @@
-package com.example.teumteum.data.remote.onboarding.dto
+package com.example.teumteum.data.remote.onboarding.model
 
 import com.google.gson.annotations.SerializedName
