+++ app/src/main/java/com/example/teumteum/ui/signup/OnBoardingSleepPatternFragment.kt
@@ -13,7 +13,7 @@ import android.widget.TextView
 import android.widget.Toast
 import com.example.teumteum.R
 import com.example.teumteum.data.remote.onboarding.OnBoardingService
-import com.example.teumteum.data.remote.onboarding.dto.SleepPatternRequest
+import com.example.teumteum.data.remote.onboarding.model.SleepPatternRequest
 import com.example.teumteum.databinding.FragmentOnBoardingSleepPatternBinding
 import com.example.teumteum.ui.signup.view.SleepPatternView
 import com.google.android.material.bottomsheet.BottomSheetDialog
